Adding a database index
Added the 'indexes' option to the model's Meta class. This option allows defining database indexes that can include one or more fields in ascending or descending order, or functional expressions and database functions. An index was added to the 'publish' field with a hyphen to define it in descending order.