Adding a status field
Defined the Status enumeration class by subclassing models.TextChoices. The available status options for posts are DRAFT and PUBLISHED, with corresponding values DF and PB, and labels Draft and Published. Added a new status field to the model, an instance of CharField, with the choices parameter to restrict its value to options from Status.choices.