Creating and applying migrations
Created the file 0001_initial.py in the migrations directory of the blog app. This migration contains SQL instructions to create the database table for the Post model and define an index for the publish field. Any changes to models.py, such as adding, removing, or modifying fields, or adding new models, will require creating new migrations using the makemigrations command. Each migration allows Django to track changes to the model. Then, apply the migration using the migrate command to synchronize the database with the models.