Creating model managers
The first manager declared in a model becomes the default manager. To specify a different default manager, use the Meta attribute default_manager_name. If no manager is defined in the model, Django automatically creates a standard manager named objects. In the provided code, the standard manager objects and a custom manager named published were added to the Post model.

The get_queryset() method of the manager returns a QuerySet that will be executed. This method was overridden to create a custom QuerySet that filters posts by their status, returning only posts with the status PUBLISHED.