Make post_list view
The post_list view takes an object request as its only parameter, which is required for all view functions. This view retrieves all posts with the status PUBLISHED using the previously created published manager. The Django shortcut function render() was used to render the list of posts with the specified template.