Make post_detail view
This view provides detailed information about a post. The view takes the post's id as an argument. Here, we retrieve the Post object with the given id by calling the get() method of the standard manager objects. We raise an Http404 exception to return a 404 HTTP error if a DoesNotExist exception occurs, meaning the model does not exist because the result is not found.